refactor: create hashOperators
