+++ src/operstors/basicOperation.mjs
@@ -53,18 +53,4 @@ export const fileOperator = {
         const filePath = `${FileManager.currentDirectory}/${path}`;
         await fsPromises.unlink(filePath);
     },
-
-    hash : async (path) => {
-      const filePath = `${FileManager.currentDirectory}/${path}`;
-      const hash = createHash('sha256');
-      const readStream = createReadStream(filePath);
-  
-      readStream.on('data', (data) => hash.update(data));
-      await new Promise((resolve, reject) => {
-        readStream.on('end', resolve);
-        readStream.on('error', reject);
-      });
-  
-      console.log(`File hash: ${hash.digest('hex')}`);
-    }
 }
